# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Documentation",
        "description": "Store API Postman and Code definitions documentation.",
        "contact": {
            "name": "Backend Engineer"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/categories/{slug}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get Category",
                "description": "Get a single category by slug",
                "operationId": "6b5e99ab9669011f1260b2a8fb93392e",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "The slug of the category",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotFoundResponse": {
                "title": "NotFound Response",
                "description": "Response for 404 not found error",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Resource not found."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "Category": {
                "title": "Category",
                "description": "Product category (detail view)",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Electronics"
                    },
                    "slug": {
                        "type": "string",
                        "example": "electronics"
                    },
                    "meta_description": {
                        "type": "string",
                        "nullable": true
                    },
                    "image": {
                        "properties": {
                            "thumb": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "card": {
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            },
                            "banner": {
                                "description": "Only in category detail endpoint",
                                "type": "string",
                                "format": "uri",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "NotFoundResponse": {
                "description": "Not found error response",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFoundResponse"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}